Moves matlab setup to the settings dialog
